refactor: 예약 검증 책임 분리
