+++ src/main/java/roomescape/service/ReservationValidator.java
@@ -1,0 +1,76 @@
+package roomescape.service;
+
+import org.springframework.stereotype.Component;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationTime;
+import roomescape.exception.DuplicateReservationException;
+import roomescape.exception.ForbiddenReservationException;
+import roomescape.exception.InvalidInputException;
+import roomescape.exception.PastReservationException;
+import roomescape.exception.PastReservationLockedException;
+import roomescape.exception.UnchangedReservationException;
+import roomescape.repository.ReservationRepository;
+
+import java.time.LocalDate;
+import java.time.LocalDateTime;
+
+@Component
+public class ReservationValidator {
+
+    private final ReservationRepository reservationRepository;
+
+    public ReservationValidator(ReservationRepository reservationRepository) {
+        this.reservationRepository = reservationRepository;
+    }
+
+    public void validateNotPast(LocalDate date, ReservationTime time) {
+        LocalDateTime reservationDateTime = LocalDateTime.of(date, time.getStartAt());
+        if (reservationDateTime.isBefore(LocalDateTime.now())) {
+            throw new PastReservationException("이미 지난 시간으로는 예약할 수 없습니다.");
+        }
+    }
+
+    public void validateAlreadyReserved(LocalDate date, Long timeId, Long themeId) {
+        if (reservationRepository.existsWith(date, timeId, themeId)) {
+            throw new DuplicateReservationException("이미 예약된 시간입니다.");
+        }
+    }
+
+    public void validateUpdatableReservation(Reservation reservation, String name) {
+        validateOwner(reservation, name);
+        validateReservationNotLocked(reservation);
+    }
+
+    public void validateUpdatePolicy(Reservation reservation, Reservation updatedReservation) {
+        validateScheduleChanged(reservation, updatedReservation);
+        validateNotPast(updatedReservation.getDate(), updatedReservation.getTime());
+        validateAlreadyReserved(
+                updatedReservation.getDate(),
+                updatedReservation.getTime().getId(),
+                updatedReservation.getTheme().getId());
+    }
+
+    public void validateUpdateValueExists(LocalDate date, Long timeId) {
+        if (date == null && timeId == null) {
+            throw new InvalidInputException("변경할 날짜 또는 시간이 필요합니다.");
+        }
+    }
+
+    private void validateOwner(Reservation reservation, String name) {
+        if (!reservation.isOwnedBy(name)) {
+            throw new ForbiddenReservationException("본인의 예약만 변경하거나 취소할 수 있습니다.");
+        }
+    }
+
+    private void validateReservationNotLocked(Reservation reservation) {
+        if (reservation.isPast()) {
+            throw new PastReservationLockedException("이미 지난 예약은 변경하거나 취소할 수 없습니다.");
+        }
+    }
+
+    private void validateScheduleChanged(Reservation reservation, Reservation updatedReservation) {
+        if (reservation.hasSameSchedule(updatedReservation)) {
+            throw new UnchangedReservationException("기존 예약과 같은 날짜·시간으로는 변경할 수 없습니다.");
+        }
+    }
+}
