+++ src/test/java/roomescape/service/ReservationServiceTest.java
@@ -1,36 +1,28 @@
 package roomescape.service;
 
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.ArgumentCaptor;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
 import roomescape.domain.Theme;
-import roomescape.exception.DuplicateReservationException;
-import roomescape.exception.ForbiddenReservationException;
 import roomescape.exception.InvalidInputException;
 import roomescape.exception.NotFoundException;
 import roomescape.exception.PastReservationException;
-import roomescape.exception.PastReservationLockedException;
-import roomescape.exception.UnchangedReservationException;
 import roomescape.repository.ReservationRepository;
+import roomescape.repository.ReservationTimeRepository;
+import roomescape.repository.ThemeRepository;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
@@ -38,10 +30,14 @@ import static org.mockito.Mockito.when;
 class ReservationServiceTest {
 
     private final ReservationRepository reservationRepository = mock();
-    private final ReservationCreator reservationCreator = mock();
+    private final ReservationTimeRepository reservationTimeRepository = mock();
+    private final ThemeRepository themeRepository = mock();
+    private final ReservationValidator reservationValidator = new ReservationValidator(reservationRepository);
     private final ReservationService service = new ReservationService(
             reservationRepository,
-            reservationCreator);
+            reservationTimeRepository,
+            themeRepository,
+            reservationValidator);
 
     private final LocalDate date = LocalDate.now().plusDays(1);
 
@@ -65,44 +61,73 @@ class ReservationServiceTest {
         verify(reservationRepository).findByName(name);
     }
 
-    @ParameterizedTest
-    @MethodSource("pastReservationDateTimes")
-    void 사용자는_지난_날짜나_시간으로_예약_생성시_예외_발생(LocalDate date, LocalTime startAt) {
-        // given
-        Long timeId = 1L;
-        Long themeId = 1L;
-        ReservationTime time = new ReservationTime(timeId, startAt);
-        when(reservationCreator.findReservationTime(timeId))
-                .thenReturn(time);
-
-        // when & then
-        assertThatThrownBy(() -> service.create("브라운", date, timeId, themeId))
-                .isInstanceOf(PastReservationException.class)
-                .hasMessage("이미 지난 시간으로는 예약할 수 없습니다.");
-
-        verify(reservationCreator).findReservationTime(timeId);
-        verify(reservationCreator, never()).create("브라운", date, timeId, themeId);
-        verifyNoInteractions(reservationRepository);
-    }
-
-    private static Stream<Arguments> pastReservationDateTimes() {
-        return Stream.of(
-                Arguments.of(LocalDate.now().minusDays(1), LocalTime.parse("08:00")),
-                Arguments.of(LocalDate.now(), LocalTime.MIDNIGHT)
-        );
-    }
-
     @Test
-    void 사용자_본인_예약_삭제_테스트() {
+    void 사용자_예약을_생성한다() {
         // given
         Long id = 1L;
         String name = "브라운";
-        Reservation reservation = new Reservation(
-                id,
-                name,
-                date,
-                new ReservationTime(1L, LocalTime.parse("08:00")),
-                new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소"));
+        Long timeId = 1L;
+        Long themeId = 1L;
+        ReservationTime time = new ReservationTime(timeId, LocalTime.parse("08:00"));
+        Theme theme = new Theme(themeId, "테스트 테마", "테마 설명", "썸네일 주소");
+        Reservation savedReservation = new Reservation(id, name, date, time, theme);
+        when(reservationTimeRepository.findBy(timeId))
+                .thenReturn(Optional.of(time));
+        when(reservationRepository.existsWith(date, timeId, themeId))
+                .thenReturn(false);
+        when(themeRepository.findBy(themeId))
+                .thenReturn(Optional.of(theme));
+        when(reservationRepository.insert(any(Reservation.class)))
+                .thenReturn(id);
+        when(reservationRepository.findById(id))
+                .thenReturn(Optional.of(savedReservation));
+
+        // when
+        Reservation result = service.create(name, date, timeId, themeId);
+
+        // then
+        ArgumentCaptor<Reservation> captor = ArgumentCaptor.forClass(Reservation.class);
+        assertThat(result).isEqualTo(savedReservation);
+        verify(reservationTimeRepository).findBy(timeId);
+        verify(reservationRepository).existsWith(date, timeId, themeId);
+        verify(themeRepository).findBy(themeId);
+        verify(reservationRepository).insert(captor.capture());
+        Reservation captured = captor.getValue();
+        assertAll(
+                () -> assertThat(captured.getId()).isNull(),
+                () -> assertThat(captured.getName()).isEqualTo(name),
+                () -> assertThat(captured.getDate()).isEqualTo(date),
+                () -> assertThat(captured.getTime()).isEqualTo(time),
+                () -> assertThat(captured.getTheme()).isEqualTo(theme));
+    }
+
+    @Test
+    void 사용자는_지난_날짜나_시간으로_예약_생성시_예외_발생() {
+        // given
+        Long timeId = 1L;
+        Long themeId = 1L;
+        LocalDate pastDate = LocalDate.now().minusDays(1);
+        ReservationTime time = new ReservationTime(timeId, LocalTime.parse("08:00"));
+        when(reservationTimeRepository.findBy(timeId))
+                .thenReturn(Optional.of(time));
+
+        // when & then
+        assertThatThrownBy(() -> service.create("브라운", pastDate, timeId, themeId))
+                .isInstanceOf(PastReservationException.class)
+                .hasMessage("이미 지난 시간으로는 예약할 수 없습니다.");
+
+        verify(reservationTimeRepository).findBy(timeId);
+        verify(reservationRepository, never()).existsWith(pastDate, timeId, themeId);
+        verifyNoInteractions(themeRepository);
+        verify(reservationRepository, never()).insert(any(Reservation.class));
+    }
+
+    @Test
+    void 사용자_본인_예약을_삭제한다() {
+        // given
+        Long id = 1L;
+        String name = "브라운";
+        Reservation reservation = createReservation(id, name, date, new ReservationTime(1L, LocalTime.parse("08:00")));
         when(reservationRepository.findById(id))
                 .thenReturn(Optional.of(reservation));
 
@@ -115,70 +140,7 @@ class ReservationServiceTest {
     }
 
     @Test
-    void 존재하지_않는_예약_삭제시_예외_발생() {
-        // given
-        Long id = 999L;
-        String name = "브라운";
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.empty());
-
-        // when & then
-        assertThatThrownBy(() -> service.delete(id, name))
-                .isInstanceOf(NotFoundException.class)
-                .hasMessage("존재하지 않는 예약입니다.");
-
-        verify(reservationRepository).findById(id);
-        verify(reservationRepository, never()).delete(id);
-    }
-
-    @Test
-    void 본인의_예약이_아니면_삭제시_예외_발생() {
-        // given
-        Long id = 1L;
-        String name = "브라운";
-        Reservation reservation = new Reservation(
-                id,
-                "구구",
-                date,
-                new ReservationTime(1L, LocalTime.parse("08:00")),
-                new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소"));
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.of(reservation));
-
-        // when & then
-        assertThatThrownBy(() -> service.delete(id, name))
-                .isInstanceOf(ForbiddenReservationException.class)
-                .hasMessage("본인의 예약만 변경하거나 취소할 수 있습니다.");
-
-        verify(reservationRepository).findById(id);
-        verify(reservationRepository, never()).delete(id);
-    }
-
-    @Test
-    void 이미_지난_예약_삭제시_예외_발생() {
-        // given
-        Long id = 1L;
-        String name = "브라운";
-        Reservation reservation = new Reservation(
-                id,
-                name,
-                LocalDate.now().minusDays(1),
-                new ReservationTime(1L, LocalTime.parse("08:00")),
-                new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소"));
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.of(reservation));
-
-        // when & then
-        assertThatThrownBy(() -> service.delete(id, name))
-                .isInstanceOf(PastReservationLockedException.class)
-                .hasMessage("이미 지난 예약은 변경하거나 취소할 수 없습니다.");
-
-        verify(reservationRepository).findById(id);
-        verify(reservationRepository, never()).delete(id);
-    }
-
-    @Test
-    void 사용자_본인_예약_변경_테스트() {
+    void 사용자_본인_예약을_변경한다() {
         // given
         Long id = 1L;
         String name = "브라운";
@@ -186,27 +148,21 @@ class ReservationServiceTest {
         LocalDate updateDate = date.plusDays(1);
         ReservationTime originalTime = new ReservationTime(1L, LocalTime.parse("08:00"));
         ReservationTime updateTime = new ReservationTime(timeId, LocalTime.parse("10:00"));
-        Theme theme = new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소");
-        Reservation reservation = new Reservation(id, name, date, originalTime, theme);
-        Reservation updatedReservation = new Reservation(id, name, updateDate, updateTime, theme);
+        Reservation reservation = createReservation(id, name, date, originalTime);
+        Reservation updatedReservation = createReservation(id, name, updateDate, updateTime);
         when(reservationRepository.findById(id))
                 .thenReturn(Optional.of(reservation), Optional.of(updatedReservation));
-        when(reservationCreator.findReservationTime(timeId))
-                .thenReturn(updateTime);
+        when(reservationTimeRepository.findBy(timeId))
+                .thenReturn(Optional.of(updateTime));
+        when(reservationRepository.existsWith(updateDate, timeId, reservation.getTheme().getId()))
+                .thenReturn(false);
 
         // when
         Reservation result = service.update(id, name, updateDate, timeId);
 
         // then
-        assertAll(
-                () -> assertThat(result.getId()).isEqualTo(id),
-                () -> assertThat(result.getDate()).isEqualTo(updateDate),
-                () -> assertThat(result.getTime()).isEqualTo(updateTime),
-                () -> assertThat(result.getTheme()).isEqualTo(theme));
         ArgumentCaptor<Reservation> captor = ArgumentCaptor.forClass(Reservation.class);
-        verify(reservationRepository, times(2)).findById(id);
-        verify(reservationCreator).findReservationTime(timeId);
-        verify(reservationRepository).existsWith(updateDate, timeId, theme.getId());
+        assertThat(result).isEqualTo(updatedReservation);
         verify(reservationRepository).update(captor.capture());
         Reservation captured = captor.getValue();
         assertAll(
@@ -214,7 +170,7 @@ class ReservationServiceTest {
                 () -> assertThat(captured.getName()).isEqualTo(name),
                 () -> assertThat(captured.getDate()).isEqualTo(updateDate),
                 () -> assertThat(captured.getTime()).isEqualTo(updateTime),
-                () -> assertThat(captured.getTheme()).isEqualTo(theme));
+                () -> assertThat(captured.getTheme()).isEqualTo(reservation.getTheme()));
     }
 
     @Test
@@ -224,11 +180,12 @@ class ReservationServiceTest {
         String name = "브라운";
         LocalDate updateDate = date.plusDays(1);
         ReservationTime time = new ReservationTime(1L, LocalTime.parse("08:00"));
-        Theme theme = new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소");
-        Reservation reservation = new Reservation(id, name, date, time, theme);
-        Reservation updatedReservation = new Reservation(id, name, updateDate, time, theme);
+        Reservation reservation = createReservation(id, name, date, time);
+        Reservation updatedReservation = createReservation(id, name, updateDate, time);
         when(reservationRepository.findById(id))
                 .thenReturn(Optional.of(reservation), Optional.of(updatedReservation));
+        when(reservationRepository.existsWith(updateDate, time.getId(), reservation.getTheme().getId()))
+                .thenReturn(false);
 
         // when
         Reservation result = service.update(id, name, updateDate, null);
@@ -237,47 +194,8 @@ class ReservationServiceTest {
         assertAll(
                 () -> assertThat(result.getDate()).isEqualTo(updateDate),
                 () -> assertThat(result.getTime()).isEqualTo(time));
-        ArgumentCaptor<Reservation> captor = ArgumentCaptor.forClass(Reservation.class);
-        verifyNoInteractions(reservationCreator);
-        verify(reservationRepository).existsWith(updateDate, time.getId(), theme.getId());
-        verify(reservationRepository).update(captor.capture());
-        Reservation captured = captor.getValue();
-        assertAll(
-                () -> assertThat(captured.getDate()).isEqualTo(updateDate),
-                () -> assertThat(captured.getTime()).isEqualTo(time));
-    }
-
-    @Test
-    void 사용자_본인_예약_변경시_시간만_변경한다() {
-        // given
-        Long id = 1L;
-        String name = "브라운";
-        Long timeId = 2L;
-        ReservationTime originalTime = new ReservationTime(1L, LocalTime.parse("08:00"));
-        ReservationTime updateTime = new ReservationTime(timeId, LocalTime.parse("10:00"));
-        Theme theme = new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소");
-        Reservation reservation = new Reservation(id, name, date, originalTime, theme);
-        Reservation updatedReservation = new Reservation(id, name, date, updateTime, theme);
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.of(reservation), Optional.of(updatedReservation));
-        when(reservationCreator.findReservationTime(timeId))
-                .thenReturn(updateTime);
-
-        // when
-        Reservation result = service.update(id, name, null, timeId);
-
-        // then
-        assertAll(
-                () -> assertThat(result.getDate()).isEqualTo(date),
-                () -> assertThat(result.getTime()).isEqualTo(updateTime));
-        ArgumentCaptor<Reservation> captor = ArgumentCaptor.forClass(Reservation.class);
-        verify(reservationCreator).findReservationTime(timeId);
-        verify(reservationRepository).existsWith(date, timeId, theme.getId());
-        verify(reservationRepository).update(captor.capture());
-        Reservation captured = captor.getValue();
-        assertAll(
-                () -> assertThat(captured.getDate()).isEqualTo(date),
-                () -> assertThat(captured.getTime()).isEqualTo(updateTime));
+        verifyNoInteractions(reservationTimeRepository);
+        verify(reservationRepository).update(any(Reservation.class));
     }
 
     @Test
@@ -285,12 +203,7 @@ class ReservationServiceTest {
         // given
         Long id = 1L;
         String name = "브라운";
-        Reservation reservation = new Reservation(
-                id,
-                name,
-                date,
-                new ReservationTime(1L, LocalTime.parse("08:00")),
-                new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소"));
+        Reservation reservation = createReservation(id, name, date, new ReservationTime(1L, LocalTime.parse("08:00")));
         when(reservationRepository.findById(id))
                 .thenReturn(Optional.of(reservation));
 
@@ -300,61 +213,7 @@ class ReservationServiceTest {
                 .hasMessage("변경할 날짜 또는 시간이 필요합니다.");
 
         verify(reservationRepository).findById(id);
-        verifyNoInteractions(reservationCreator);
-        verify(reservationRepository, never()).update(any(Reservation.class));
-    }
-
-    @Test
-    void 기존_날짜와_시간으로_예약_변경시_예외_발생() {
-        // given
-        Long id = 1L;
-        String name = "브라운";
-        Long timeId = 1L;
-        ReservationTime time = new ReservationTime(timeId, LocalTime.parse("08:00"));
-        Theme theme = new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소");
-        Reservation reservation = new Reservation(id, name, date, time, theme);
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.of(reservation));
-        when(reservationCreator.findReservationTime(timeId))
-                .thenReturn(time);
-
-        // when & then
-        assertThatThrownBy(() -> service.update(id, name, date, timeId))
-                .isInstanceOf(UnchangedReservationException.class)
-                .hasMessage("기존 예약과 같은 날짜·시간으로는 변경할 수 없습니다.");
-
-        verify(reservationRepository).findById(id);
-        verify(reservationCreator).findReservationTime(timeId);
-        verify(reservationRepository, never()).existsWith(any(LocalDate.class), anyLong(), anyLong());
-        verify(reservationRepository, never()).update(any(Reservation.class));
-    }
-
-    @Test
-    void 이미_예약된_시간으로_예약_변경시_예외_발생() {
-        // given
-        Long id = 1L;
-        String name = "브라운";
-        Long timeId = 2L;
-        LocalDate updateDate = date.plusDays(1);
-        ReservationTime originalTime = new ReservationTime(1L, LocalTime.parse("08:00"));
-        ReservationTime updateTime = new ReservationTime(timeId, LocalTime.parse("10:00"));
-        Theme theme = new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소");
-        Reservation reservation = new Reservation(id, name, date, originalTime, theme);
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.of(reservation));
-        when(reservationCreator.findReservationTime(timeId))
-                .thenReturn(updateTime);
-        when(reservationRepository.existsWith(updateDate, timeId, theme.getId()))
-                .thenReturn(true);
-
-        // when & then
-        assertThatThrownBy(() -> service.update(id, name, updateDate, timeId))
-                .isInstanceOf(DuplicateReservationException.class)
-                .hasMessage("이미 예약된 시간입니다.");
-
-        verify(reservationRepository).findById(id);
-        verify(reservationCreator).findReservationTime(timeId);
-        verify(reservationRepository).existsWith(updateDate, timeId, theme.getId());
+        verifyNoInteractions(reservationTimeRepository);
         verify(reservationRepository, never()).update(any(Reservation.class));
     }
 
@@ -362,45 +221,16 @@ class ReservationServiceTest {
     void 존재하지_않는_예약_변경시_예외_발생() {
         // given
         Long id = 999L;
-        String name = "브라운";
-        Long timeId = 2L;
-        LocalDate updateDate = date.plusDays(1);
         when(reservationRepository.findById(id))
                 .thenReturn(Optional.empty());
 
         // when & then
-        assertThatThrownBy(() -> service.update(id, name, updateDate, timeId))
+        assertThatThrownBy(() -> service.update(id, "브라운", date, 1L))
                 .isInstanceOf(NotFoundException.class)
                 .hasMessage("존재하지 않는 예약입니다.");
 
         verify(reservationRepository).findById(id);
-        verifyNoInteractions(reservationCreator);
-        verify(reservationRepository, never()).update(any(Reservation.class));
-    }
-
-    @Test
-    void 본인의_예약이_아니면_변경시_예외_발생() {
-        // given
-        Long id = 1L;
-        String name = "브라운";
-        Long timeId = 2L;
-        LocalDate updateDate = date.plusDays(1);
-        Reservation reservation = new Reservation(
-                id,
-                "구구",
-                date,
-                new ReservationTime(1L, LocalTime.parse("08:00")),
-                new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소"));
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.of(reservation));
-
-        // when & then
-        assertThatThrownBy(() -> service.update(id, name, updateDate, timeId))
-                .isInstanceOf(ForbiddenReservationException.class)
-                .hasMessage("본인의 예약만 변경하거나 취소할 수 있습니다.");
-
-        verify(reservationRepository).findById(id);
-        verifyNoInteractions(reservationCreator);
+        verifyNoInteractions(reservationTimeRepository);
         verify(reservationRepository, never()).update(any(Reservation.class));
     }
 
@@ -410,81 +240,28 @@ class ReservationServiceTest {
         Long id = 1L;
         String name = "브라운";
         Long timeId = 999L;
-        LocalDate updateDate = date.plusDays(1);
-        Reservation reservation = new Reservation(
-                id,
-                name,
-                date,
-                new ReservationTime(1L, LocalTime.parse("08:00")),
-                new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소"));
+        Reservation reservation = createReservation(id, name, date, new ReservationTime(1L, LocalTime.parse("08:00")));
         when(reservationRepository.findById(id))
                 .thenReturn(Optional.of(reservation));
-        when(reservationCreator.findReservationTime(timeId))
-                .thenThrow(new NotFoundException("존재하지 않는 예약 시간입니다."));
+        when(reservationTimeRepository.findBy(timeId))
+                .thenReturn(Optional.empty());
 
         // when & then
-        assertThatThrownBy(() -> service.update(id, name, updateDate, timeId))
+        assertThatThrownBy(() -> service.update(id, name, date.plusDays(1), timeId))
                 .isInstanceOf(NotFoundException.class)
                 .hasMessage("존재하지 않는 예약 시간입니다.");
 
         verify(reservationRepository).findById(id);
-        verify(reservationCreator).findReservationTime(timeId);
+        verify(reservationTimeRepository).findBy(timeId);
         verify(reservationRepository, never()).update(any(Reservation.class));
     }
 
-    @Test
-    void 이미_지난_예약_변경시_예외_발생() {
-        // given
-        Long id = 1L;
-        String name = "브라운";
-        Long timeId = 2L;
-        LocalDate updateDate = date.plusDays(1);
-        Reservation reservation = new Reservation(
-                id,
-                name,
-                LocalDate.now().minusDays(1),
-                new ReservationTime(1L, LocalTime.parse("08:00")),
-                new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소"));
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.of(reservation));
-
-        // when & then
-        assertThatThrownBy(() -> service.update(id, name, updateDate, timeId))
-                .isInstanceOf(PastReservationLockedException.class)
-                .hasMessage("이미 지난 예약은 변경하거나 취소할 수 없습니다.");
-
-        verify(reservationRepository).findById(id);
-        verifyNoInteractions(reservationCreator);
-        verify(reservationRepository, never()).update(any(Reservation.class));
-    }
-
-    @Test
-    void 지난_날짜와_시간으로_예약_변경시_예외_발생() {
-        // given
-        Long id = 1L;
-        String name = "브라운";
-        Long timeId = 2L;
-        LocalDate updateDate = LocalDate.now().minusDays(1);
-        Reservation reservation = new Reservation(
+    private Reservation createReservation(Long id, String name, LocalDate date, ReservationTime time) {
+        return new Reservation(
                 id,
                 name,
                 date,
-                new ReservationTime(1L, LocalTime.parse("08:00")),
+                time,
                 new Theme(1L, "테스트 테마", "테마 설명", "썸네일 주소"));
-        ReservationTime updateTime = new ReservationTime(timeId, LocalTime.parse("10:00"));
-        when(reservationRepository.findById(id))
-                .thenReturn(Optional.of(reservation));
-        when(reservationCreator.findReservationTime(timeId))
-                .thenReturn(updateTime);
-
-        // when & then
-        assertThatThrownBy(() -> service.update(id, name, updateDate, timeId))
-                .isInstanceOf(PastReservationException.class)
-                .hasMessage("이미 지난 시간으로는 예약할 수 없습니다.");
-
-        verify(reservationRepository).findById(id);
-        verify(reservationCreator).findReservationTime(timeId);
-        verify(reservationRepository, never()).update(any(Reservation.class));
     }
-
 }
